requirements.txt
flask
google-genai
chromadb
PyPDF2
python-dotenv

